Makefile -> CMake, use boost primarily
