+++ CMakeLists.txt
@@ -1,0 +1,5 @@
+
+cmake_minimum_required (VERSION 2.8.11)
+project (grip)
+
+add_subdirectory (src)
