+++ src/CMakeLists.txt
@@ -1,0 +1,5 @@
+add_subdirectory (external)
+add_subdirectory (general)
+add_subdirectory (grip)
+add_subdirectory (gripgen)
+add_subdirectory (test)
