+++ src/grip/CMakeLists.txt
@@ -1,0 +1,26 @@
+find_package(Boost REQUIRED COMPONENTS regex filesystem system)
+if(Boost_FOUND)
+    include_directories(${Boost_INCLUDE_DIRS})
+    if(MINGW)
+        set(CMAKE_EXE_LINKER_FLAGS "-static -static-libgcc -static-libstdc++")
+    endif()
+    add_library (Common grep.cpp glob.cpp pattern.cpp)
+    target_link_libraries(Common LINK_PUBLIC ${Boost_LIBRARIES} External General)
+    add_executable (egrip egrip.cpp)
+    add_executable (fgrip fgrip.cpp)
+    add_executable (grip grip.cpp)
+    if(MINGW)
+        set_target_properties(egrip PROPERTIES LINK_SEARCH_START_STATIC 1)
+        set_target_properties(egrip PROPERTIES LINK_SEARCH_END_STATIC 1)
+        set_target_properties(fgrip PROPERTIES LINK_SEARCH_START_STATIC 1)
+        set_target_properties(fgrip PROPERTIES LINK_SEARCH_END_STATIC 1)
+        set_target_properties(grip PROPERTIES LINK_SEARCH_START_STATIC 1)
+        set_target_properties(grip PROPERTIES LINK_SEARCH_END_STATIC 1)
+        set(CMAKE_FIND_LIBRARY_SUFFIXES ".a")
+    endif()
+    target_link_libraries(egrip LINK_PUBLIC ${Boost_LIBRARIES} Common)
+    target_link_libraries(fgrip LINK_PUBLIC ${Boost_LIBRARIES} Common)
+    target_link_libraries(grip LINK_PUBLIC ${Boost_LIBRARIES} Common)
+else()
+    message("Boost libraries (regex filesystem system) not found!")
+endif()
