+++ src/gripgen/gripgen.cpp
@@ -10,14 +10,11 @@
 #include <cstdlib>
 #include <climits>
 
-#ifdef _POSIX_C_SOURCE
-#include <unistd.h>
-#include <getopt.h>
-#else
-#include "external/getopt.h"
-#include "external/getopt.c"
-#include "external/getopt1.c"
-#endif
+extern "C" {
+#include "getopt.h"
+#include "getopt.c"
+#include "getopt1.c"
+}
 
 using namespace std;
 using namespace std::chrono;
