--- src/scripts/install.cmd
@@ -1,10 +1,0 @@
-@echo off
-for %%a in (%*) do set "dst=%%a"
-set "dst=%dst:/=\%"
-
-:loop
-	set "src=%1"
-	set "src=%src:/=\%"
-	copy "%src%" %dst%"
-	shift
-if not "%~2"=="" goto loop
