--- src/scripts/mkdir.cmd
@@ -1,7 +1,0 @@
-@echo off
-:loop
-	set "p=%1"
-	set "p=%p:/=\%"
-	if not exist "%p%" mkdir "%p%"
-	shift
-if not "%~1"=="" goto loop
