+++ src/test/CMakeLists.txt
@@ -1,0 +1,16 @@
+find_package(Boost REQUIRED COMPONENTS regex filesystem system)
+if(Boost_FOUND)
+    include_directories(${Boost_INCLUDE_DIRS})
+    if(MINGW)
+        set(CMAKE_EXE_LINKER_FLAGS "-static -static-libgcc -static-libstdc++")
+    endif()
+    add_executable (test test.cpp ids.cpp compressedids.cpp pattern.cpp ../grip/pattern.cpp)
+    if(MINGW)
+        set_target_properties(test PROPERTIES LINK_SEARCH_START_STATIC 1)
+        set_target_properties(test PROPERTIES LINK_SEARCH_END_STATIC 1)
+        set(CMAKE_FIND_LIBRARY_SUFFIXES ".a")
+    endif()
+    target_link_libraries (test LINK_PUBLIC ${Boost_LIBRARIES} External General)
+else()
+    message("Boost libraries (regex filesystem system) not found!")
+endif()
