+++ src/test/test.cpp
@@ -1,3 +1,3 @@
 #define CATCH_CONFIG_MAIN
-#include "external/catch.hpp"
+#include "catch.hpp"
